refactor the purchase credit flow to components
