+++ dashboard/src/hooks/useWallet.ts
@@ -1,7 +1,8 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { config } from "@/config/walletConfig";
-import { getChainId } from "@wagmi/core";
+import { getChainId, readContract } from "@wagmi/core";
 import { useCallback, useMemo } from "react";
+import { erc20Abi } from "viem";
 import { useAccount, useSwitchChain } from "wagmi";
 import { getBalance } from "wagmi/actions";
 
@@ -42,7 +43,27 @@ export default function useWallet() {
       });
       return balance.formatted;
     },
-    [address, chainId]
+    [address, chainId],
+  );
+
+  const getERC20AvailBalance = useCallback(
+    async (address: `0x${string}`) => {
+      await readContract(config, {
+        address: "0x8B42845d23C68B845e262dC3e5cAA1c9ce9eDB44" as `0x${string}`,
+        abi: erc20Abi,
+        functionName: "balanceOf",
+        args: [address],
+        chainId: activeNetworkId,
+      })
+        .then((balance) => {
+          if (!balance) return new BigNumber(0);
+          return new BigNumber(balance as bigint);
+        })
+        .catch((error) => {
+          console.log(error);
+        });
+    },
+    [activeNetworkId],
   );
 
   /**
@@ -56,5 +77,6 @@ export default function useWallet() {
     activeNetworkId,
     switchNetwork,
     showBalance,
+    getERC20AvailBalance,
   };
 }
