+++ dashboard/src/module/purchase-credit/index.tsx
@@ -1,47 +1,28 @@
 "use client";
-import Button from "@/components/button";
 import CreditsTransactionProgress from "@/components/credits-transaction-progress";
-import { useDialog } from "@/components/dialog/provider";
 import Input from "@/components/input";
 import { Text } from "@/components/text";
-import { useAppToast } from "@/components/toast";
 import { Card, CardContent } from "@/components/ui/card";
 import { config } from "@/config/walletConfig";
 import { useDebounce } from "@/hooks/useDebounce";
-import { useDesiredChain } from "@/hooks/useDesiredChain";
 import useWallet from "@/hooks/useWallet";
 import { TOKEN_MAP } from "@/lib/types";
-import { formatDataBytes, numberToBytes32 } from "@/lib/utils";
+import { formatDataBytes } from "@/lib/utils";
 import SelectTokenButton from "@/module/purchase-credit/select-token-button";
-import { TransactionStatus, useConfig } from "@/providers/ConfigProvider";
+import { useConfig } from "@/providers/ConfigProvider";
 import CreditService from "@/services/credit";
-import { SignedIn, SignedOut, SignInButton } from "@clerk/nextjs";
-import { readContract, writeContract } from "@wagmi/core";
-import {
-  AvailWalletConnect,
-  useAvailAccount,
-  useAvailWallet,
-} from "avail-wallet-sdk";
+import { readContract } from "@wagmi/core";
 import BigNumber from "bignumber.js";
-import { ConnectKitButton } from "connectkit";
-import { LoaderCircle, Wallet } from "lucide-react";
-import {
-  MouseEvent,
-  useCallback,
-  useDeferredValue,
-  useEffect,
-  useState,
-} from "react";
-import { v4 as uuidv4 } from "uuid";
-import { erc20Abi, parseUnits } from "viem";
+import { Wallet } from "lucide-react";
+import { useCallback, useDeferredValue, useEffect, useState } from "react";
+import { erc20Abi } from "viem";
 import { useAccount, useBalance as useWagmiBalance } from "wagmi";
-import { batchTransferAndRemark, postOrder } from "./utils";
-import { depositAbi } from "./utils/constant";
+import BuySection from "./components/buy-section";
 
 const DESIRED_CHAIN = 11155111;
 
 const BuyCreditsCard = ({ token }: { token?: string }) => {
-  const { activeNetworkId, showBalance } = useWallet();
+  const { getERC20AvailBalance, showBalance } = useWallet();
   const [tokenAmount, setTokenAmount] = useState("");
   const [tokenAmountError, setTokenAmountError] = useState("");
   const [estimateData, setEstimateData] = useState();
@@ -50,26 +31,15 @@ const BuyCreditsCard = ({ token }: { token?: string }) => {
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState("");
   const account = useAccount();
-  const { selected } = useAvailAccount();
-  const { api } = useAvailWallet();
-  const { setOpen } = useDialog();
-  const { error: errorToast } = useAppToast();
-  const {
-    selectedChain,
-    selectedToken,
-    setTransactionStatusList,
-    setShowTransaction,
-    availNativeBalance,
-  } = useConfig();
+  const { selectedChain, selectedToken, availNativeBalance } = useConfig();
   const balance = useWagmiBalance({
     address: account.address,
   });
   const debouncedValue = useDebounce(deferredTokenValue, 500);
-  const { chainChangerAsync } = useDesiredChain(DESIRED_CHAIN);
 
   useEffect(() => {
     if (!account.address) return;
-    getERC20AvailBalance();
+    getERC20AvailBalance(account.address);
     showBalance({ token: account.address })
       .then((response) => {
         console.log({
@@ -116,140 +86,22 @@ const BuyCreditsCard = ({ token }: { token?: string }) => {
     }
   };
 
-  const getERC20AvailBalance = useCallback(async () => {
-    await readContract(config, {
-      address: "0x8B42845d23C68B845e262dC3e5cAA1c9ce9eDB44" as `0x${string}`,
-      abi: erc20Abi,
-      functionName: "balanceOf",
-      args: [account.address!],
-      chainId: activeNetworkId,
-    })
-      .then((balance) => {
-        if (!balance) return new BigNumber(0);
-        return new BigNumber(balance as bigint);
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  }, [account, activeNetworkId]);
-
-  const handleBuyCredits = async ({ isAvail }: { isAvail: boolean }) => {
-    if (!tokenAmount) return;
-    try {
-      setLoading(true);
-      setError("");
-
-      const tokenAddress =
-        selectedToken &&
-        TOKEN_MAP[selectedToken?.name?.toLowerCase()].token_address;
-
-      console.log(tokenAddress, "token addy");
-      const orderResponse = await postOrder({
-        token: token!,
-        chainId: activeNetworkId,
-      });
-
-      if (!orderResponse?.data) {
-        setLoading(false);
-        setError(orderResponse.message);
-        return;
-      }
-
-      if (isAvail) {
-        const txn = await batchTransferAndRemark(
-          api!,
-          selected!,
-          parseUnits(tokenAmount, 18).toString(),
-          "Buy Credits",
-        );
-
-        if (txn.isOk()) {
-          const transaction: TransactionStatus = {
-            id: uuidv4(),
-            status: "finality",
-            orderId: orderResponse.data.id as number,
-            tokenAddress: tokenAddress! as `0x${string}`,
-            tokenAmount: +tokenAmount,
-            txnHash: txn.value.txhash,
-          };
-          setTransactionStatusList((prev) => [...(prev ?? []), transaction]);
-          setShowTransaction(transaction);
-          setOpen("credit-transaction");
-          setTokenAmount("");
-          setLoading(false);
-          return;
-        }
-      }
-
-      await writeContract(config, {
-        address: tokenAddress as `0x${string}`,
-        abi: erc20Abi,
-        functionName: "approve",
-        args: [
-          process.env.NEXT_PUBLIC_ADDRESS as `0x${string}`,
-          parseUnits(tokenAmount, 18),
-        ],
-        chainId: activeNetworkId,
-      })
-        .then(async () => {
-          await writeContract(config, {
-            address: process.env.NEXT_PUBLIC_ADDRESS as `0x${string}`,
-            abi: depositAbi,
-            functionName: "depositERC20",
-            args: [
-              numberToBytes32(+orderResponse?.data?.id),
-              parseUnits(tokenAmount, 18),
-              tokenAddress,
-            ],
-            chainId: activeNetworkId,
-          })
-            .then(async (txnHash: `0x${string}`) => {
-              const transaction: TransactionStatus = {
-                id: uuidv4(),
-                status: "finality",
-                orderId: orderResponse.data.id as number,
-                tokenAddress: tokenAddress! as `0x${string}`,
-                tokenAmount: +tokenAmount,
-                txnHash,
-              };
-              setTransactionStatusList((prev) => [
-                ...(prev ?? []),
-                transaction,
-              ]);
-              setShowTransaction(transaction);
-              setOpen("credit-transaction");
-              setTokenAmount("");
-              setLoading(false);
-            })
-            .catch((err) => {
-              const message = err.message.split(".")[0];
-              if (message === "User rejected the request") {
-                errorToast?.({ label: "You rejected the request" });
-              } else {
-                errorToast?.({ label: "Transaction failed" });
-              }
-              setLoading(false);
-            });
-        })
-        .catch((err) => {
-          const message = err.message.split(".")[0];
-          if (message === "User rejected the request") {
-            errorToast?.({ label: "You rejected the request" });
-          } else {
-            errorToast?.({ label: message });
-          }
-          setLoading(false);
-        });
-    } catch (error) {
-      console.log(error);
-      setLoading(false);
-    }
+  const handleBuyStart = () => {
+    setLoading(true);
+    setError("");
   };
 
-  const handleClick = (e: MouseEvent, callback?: VoidFunction) => {
-    e.preventDefault();
-    e.stopPropagation();
-    callback?.();
+  const handleBuyComplete = () => {
+    setLoading(false);
+  };
+
+  const handleBuyError = (errorMessage: string) => {
+    setError(errorMessage);
+    setLoading(false);
+  };
+
+  const handleTokenAmountClear = () => {
+    setTokenAmount("");
   };
 
   return (
@@ -358,135 +210,16 @@ const BuyCreditsCard = ({ token }: { token?: string }) => {
                 </div>
               </div>
 
-              <div className="flex-1 flex flex-col gap-y-3 items-center p-4 justify-end">
-                <SignedIn>
-                  {!!error && (
-                    <Text variant={"error"} size={"sm"} weight={"medium"}>
-                      {error}
-                    </Text>
-                  )}
-                  {(!selectedChain || !selectedToken) && (
-                    <Button
-                      onClick={() => {
-                        setOpen("select-token");
-                      }}
-                    >
-                      Connect Wallet
-                    </Button>
-                  )}
-                  {selectedChain?.name === "Ethereum" && selectedToken && (
-                    <ConnectKitButton.Custom>
-                      {(props) => {
-                        if (!props.isConnected) {
-                          return (
-                            <Button onClick={(e) => handleClick(e, props.show)}>
-                              Connect EVM Wallet
-                            </Button>
-                          );
-                        }
-
-                        if (!props.chain || props.chain?.id !== DESIRED_CHAIN) {
-                          return (
-                            <Button onClick={() => chainChangerAsync()}>
-                              Wrong Network
-                            </Button>
-                          );
-                        }
-
-                        return (
-                          <Button
-                            onClick={() => {
-                              handleBuyCredits({
-                                isAvail: false,
-                              });
-                            }}
-                            variant={
-                              !selectedToken ||
-                              !selectedChain ||
-                              !tokenAmount ||
-                              tokenAmount === "0" ||
-                              tokenAmountError !== ""
-                                ? "disabled"
-                                : "primary"
-                            }
-                            disabled={
-                              loading ||
-                              !selectedToken ||
-                              !selectedChain ||
-                              !tokenAmount ||
-                              tokenAmount === "0" ||
-                              tokenAmountError !== ""
-                            }
-                          >
-                            {loading ? (
-                              <div className="flex gap-x-1 justify-center">
-                                <LoaderCircle
-                                  className="animate-spin"
-                                  color="#fff"
-                                  size={24}
-                                />
-                                Waiting for confirmation
-                              </div>
-                            ) : (
-                              "Buy Now"
-                            )}
-                          </Button>
-                        );
-                      }}
-                    </ConnectKitButton.Custom>
-                  )}
-                  {selectedChain?.name === "Avail" && selectedToken && (
-                    <AvailWalletConnect
-                      connectedChildren={
-                        <Button
-                          onClick={() => {
-                            handleBuyCredits({
-                              isAvail: true,
-                            });
-                          }}
-                          variant={
-                            !selectedToken ||
-                            !selectedChain ||
-                            !tokenAmount ||
-                            tokenAmount === "0" ||
-                            tokenAmountError !== ""
-                              ? "disabled"
-                              : "primary"
-                          }
-                          disabled={
-                            loading ||
-                            !selectedToken ||
-                            !selectedChain ||
-                            !tokenAmount ||
-                            tokenAmount === "0" ||
-                            tokenAmountError !== ""
-                          }
-                        >
-                          {loading ? (
-                            <div className="flex gap-x-1 justify-center">
-                              <LoaderCircle
-                                className="animate-spin"
-                                color="#fff"
-                                size={24}
-                              />
-                              Waiting for confirmation
-                            </div>
-                          ) : (
-                            "Buy Now"
-                          )}
-                        </Button>
-                      }
-                    >
-                      <Button>Connect Wallet</Button>
-                    </AvailWalletConnect>
-                  )}
-                </SignedIn>
-                <SignedOut>
-                  <SignInButton mode="modal" component="div">
-                    <Button>Sign In</Button>
-                  </SignInButton>
-                </SignedOut>
-              </div>
+              <BuySection
+                tokenAmount={tokenAmount}
+                tokenAmountError={tokenAmountError}
+                error={error}
+                onBuyStart={handleBuyStart}
+                onBuyComplete={handleBuyComplete}
+                onBuyError={handleBuyError}
+                onTokenAmountClear={handleTokenAmountClear}
+                token={token}
+              />
             </CardContent>
           </div>
         </Card>
